# INSY/Projekt/REST/swagger.yaml
# swagger.yaml

openapi: 3.0.0
info:
  title: API Dokumentation
  version: 1.0.0
paths:
  /useraccounts:
    get:
      summary: Gibt alle Benutzerkonten zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Benutzerkonten zurück.
    post:
      summary: Fügt ein neues Benutzerkonto hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                phone_number:
                  type: string
                created_at:
                  type: string
                  format: date-time
              required:
                - username
                - email
                - password
                - first_name
                - last_name
                - phone_number
                - created_at
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt das hinzugefügte Benutzerkonto zurück.
  /useraccounts/{user_id}:
    delete:
      summary: Löscht den angegebenen User
      parameters:
        - in: path
          name: user_id
          required: true
          description: Die ID des zu löschenden Users
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht User.
    put:
      summary: Benutzerkonto aktualisieren
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID des Benutzerkontos
        - name: body
          in: body
          required: true
          description: Aktualisierte Benutzerkontodaten
          schema:
            type: object
            properties:
              username:
                type: string
              email:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              phone_number:
                type: string
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example:
                user_id: 1
                username: updatedUsername
                email: updated@example.com
                first_name: UpdatedFirstName
                last_name: UpdatedLastName
                phone_number: UpdatedPhoneNumber
  /addresses:
    get:
      summary: Gibt alle Adressen zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Adressen zurück.
    post:
      summary: Fügt eine neue Adresse hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                address_type:
                  type: string
                street:
                  type: string
                city:
                  type: string
                state:
                  type: string
                postal_code:
                  type: string
                country:
                  type: string
              required:
                - user_id
                - address_type
                - street
                - city
                - state
                - postal_code
                - country
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt die hinzugefügte Adresse zurück.
  /addresses/{address_id}:
    delete:
      summary: Löscht die angegebene Adresse
      parameters:
        - in: path
          name: address_id
          required: true
          description: Die ID der zu löschenden Adresse
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht Adresse.
    put:
      summary: Adresse aktualisieren
      parameters:
        - name: address_id
          in: path
          required: true
          schema:
            type: integer
          description: ID der Adresse
        - name: body
          in: body
          required: true
          description: Aktualisierte Adressendaten
          schema:
            type: object
            properties:
              address_type:
                type: string
              street:
                type: string
              city:
                type: string
              state:
                type: string
              postal_code:
                type: string
              country:
                type: string
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example:
                address_id: 1
                user_id: 1
                address_type: updatedType
                street: updatedStreet
                city: updatedCity
                state: updatedState
                postal_code: updatedPostalCode
                country: updatedCountry
  /productcatalog:
    get:
      summary: Gibt alle Produkte im Katalog zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Produkte im Katalog zurück.
    post:
      summary: Fügt ein neues Produkt zum Katalog hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product_name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                stock_quantity:
                  type: integer
                image_url:
                  type: string
              required:
                - product_name
                - description
                - price
                - stock_quantity
                - image_url
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt das hinzugefügte Produkt im Katalog zurück.
  /productcatalog/{product_id}:
    delete:
      summary: Löscht das angegebene Produkt
      parameters:
        - in: path
          name: product_id
          required: true
          description: Die ID des zu löschenden Produkts
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht Produkt.
    put:
      summary: Produkt im Katalog aktualisieren
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
          description: ID des Produkts
        - name: body
          in: body
          required: true
          description: Aktualisierte Produktdaten
          schema:
            type: object
            properties:
              product_name:
                type: string
              description:
                type: string
              price:
                type: number
              stock_quantity:
                type: integer
              image_url:
                type: string
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example:
                product_id: 1
                product_name: updatedProductName
                description: updatedDescription
                price: 99.99
                stock_quantity: 50
                image_url: updatedImageUrl
  /shoppingcart:
    get:
      summary: Gibt alle Warenkörbe zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Warenkörbe zurück.
    post:
      summary: Fügt einen neuen Warenkorb hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                created_at:
                  type: string
                  format: date-time
              required:
                - user_id
                - created_at
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt den hinzugefügten Warenkorb zurück.
  /shoppingcart/{cart_id}:
    delete:
      summary: Löscht das angegebene Shoppingcart
      parameters:
        - in: path
          name: cart_id
          required: true
          description: Die ID des zu löschenden Shoppingcarts
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht Shoppingcart.
    put:
      summary: Warenkorb aktualisieren
      parameters:
        - name: cart_id
          in: path
          required: true
          schema:
            type: integer
          description: ID des Warenkorbs
        - name: body
          in: body
          required: true
          description: Aktualisierte Warenkorbdetails
          schema:
            type: object
            properties:
              user_id:
                type: integer
              created_at:
                type: string
                format: date-time
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example:
                cart_id: 1
                user_id: 1
                created_at: '2024-01-15T12:00:00.000Z'
  /cartitems:
    get:
      summary: Gibt alle Warenkorbpositionen zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Warenkorbpositionen zurück.
    post:
      summary: Fügt eine neue Warenkorbposition hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cart_id:
                  type: integer
                product_id:
                  type: integer
                quantity:
                  type: integer
              required:
                - cart_id
                - product_id
                - quantity
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt die hinzugefügte Warenkorbposition zurück.
  /cartitems/{cart_item_id}:
    delete:
      summary: Löscht das angegebene CartItem
      parameters:
        - in: path
          name: cart_item_id
          required: true
          description: Die ID des zu löschenden CartItems
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht CartItem.
    put:
      summary: Warenkorbposition aktualisieren
      parameters:
        - name: cart_item_id
          in: path
          required: true
          schema:
            type: integer
          description: ID der Warenkorbposition
        - name: body
          in: body
          required: true
          description: Aktualisierte Warenkorbpositionsdetails
          schema:
            type: object
            properties:
              cart_id:
                type: integer
              product_id:
                type: integer
              quantity:
                type: integer
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example:
                cart_item_id: 1
                cart_id: 1
                product_id: 1
                quantity: 2
  /orders:
    get:
      summary: Gibt alle Bestellungen zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Bestellungen zurück.
    post:
      summary: Fügt eine neue Bestellung hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                payment_method:
                  type: string
                shipping_address_id:
                  type: integer
                billing_address_id:
                  type: integer
                order_status:
                  type: string
                order_date:
                  type: string
                  format: date-time
                total_cost:
                  type: number
              required:
                - user_id
                - payment_method
                - shipping_address_id
                - billing_address_id
                - order_status
                - order_date
                - total_cost
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt die hinzugefügte Bestellung zurück.
  /orders/{order_id}:
    delete:
      summary: Löscht die angegebene Order
      parameters:
        - in: path
          name: order_id
          required: true
          description: Die ID der zu löschenden Order
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht Order.
    put:
      summary: Bestellung aktualisieren
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
          description: ID der Bestellung
        - name: body
          in: body
          required: true
          description: Aktualisierte Bestelldetails
          schema:
            type: object
            properties:
              user_id:
                type: integer
              payment_method:
                type: string
              shipping_address_id:
                type: integer
              billing_address_id:
                type: integer
              order_status:
                type: string
              order_date:
                type: string
                format: date-time
              total_cost:
                type: number
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example:
                order_id: 1
                user_id: 1
                payment_method: updatedPaymentMethod
                shipping_address_id: 1
                billing_address_id: 2
                order_status: updatedStatus
                order_date: '2024-01-15T12:00:00.000Z'
                total_cost: 150.99
  /orderitems:
    get:
      summary: Gibt alle Bestellpositionen zurück.
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt eine Liste der Bestellpositionen zurück.
    post:
      summary: Fügt eine neue Bestellposition hinzu.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                product_id:
                  type: integer
                quantity:
                  type: integer
                price_at_purchase:
                  type: number
              required:
                - order_id
                - product_id
                - quantity
                - price_at_purchase
      responses:
        200:
          description: Erfolgreiche Anfrage. Gibt die hinzugefügte Bestellposition zurück.
  /orderitems/{order_item_id}:
    delete:
      summary: Löscht das angegebene OrderItem
      parameters:
        - in: path
          name: order_item_id
          required: true
          description: Die ID des zu löschenden OrderItems
          schema:
            type: integer
      responses:
        200:
          description: Erfolgreiche Anfrage. Löscht OrderItem
    put:
      summary: Bestellposition aktualisieren
      parameters:
        - name: order_item_id
          in: path
          required: true
          schema:
            type: integer
          description: ID der Bestellposition
        - name: body
          in: body
          required: true
          description: Aktualisierte Bestellpositionsdetails
          schema:
            type: object
            properties:
              order_id:
                type: integer
              product_id:
                type: integer
              quantity:
                type: integer
              price_at_purchase:
                type: number
      responses:
        '200':
          description: Erfolgreich aktualisiert
          content:
            application/json:
              example: